requirements.txt
streamlit
google-generativeai
duckduckgo-search
python-dotenv